requirements.txt
Django==4.2.9
djangorestframework==3.14.0
pillow==10.2.0
django_extensions==3.1.3
